test: move LookupSubject scope regressions to base specs; harden SqlServer attr filter
Address PR review feedback:

- Move the LookupSubject entity-scope regression tests out of the InMemory-only
  LookupSubjectEntityScopeRepro class and into BaseLookupSubjectEngineSpecs so they
  run against every provider (InMemory, Postgres, SqlServer). The bug was in the SQL
  builders, so the coverage belongs at the provider-agnostic layer. Adds a CreateScope
  helper to resolve the writer / Check engine alongside the lookup engine from one scope.
- Drop the NSubstitute-based CascadingHop_NeverForwardsEmptyEntityIdList test: the
  invariant it asserted is already enforced by the empty-EntitiesIds guard in the engine
  and cannot break without failing the behavioral tests; mocking IDataReaderProvider
  coupled the test to engine internals.
- Add the missing empty-list "1 = 0" guard to the FilterAttributes(AttributeFilter,
  IEnumerable<string>, string) overload in SqlServer, matching the defense-in-depth
  applied to the neighboring overloads and avoiding an empty TVP.
